requirements.txt
Django==5.0.7
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.6
pandas==2.2.2
psycopg2==2.9.9 --no-binary psycopg2